New 'oocairo' module, binding Cairo library
This is just a beginning, but what's here seems to work.  I'm committing
a couple of example programs which already work fine.

There's already a library called 'luacairo' on luaforge.net, but its API
is exactly the same as the C one (in fact you could copy a lot of C code
into a Lua file and much of it would still work).  My binding is
designed to take advantage of Lua a bit more.  The 'oo' in the name
reflects the fact that I'm using method calls on objects rather than
straight functions, although really it's just a convenient namespacing
mechanism here.