oocairo: add masking and new ways to load PNG images
PNG images can now be loaded from Lua file handles (or any object with a
'read' method which accepts the number of bytes to return), and from a
Lua string.  The latter allows loading non-PNG images by loading them
first with the Lua 'gd' binding and then writing them into a string as a
PNG (as demonstrated in the new 'masking.lua' example which loads the
new example JPEG images, which would make the distribution annoyingly
big if they were in PNG form).