oocairo: support SVG surfaces, and more error checking
I've now got support for creating an SVG surface given the filename to
write the output to, and I throw exceptions when an SVG surface is used
with methods which are only meant to work on image surfaces.  Similarly,
a few gradient-only methods now throw exceptions when called on other
kinds of pattern.

Added the 'flush' and 'finish' methods for surfaces.