oocairo: add support for PDF output
This includes a 'set_size' method which only works on PDF surfaces, so
I've set it to throw an exception on any other kind of surface.

Factored out the temp-file cleanup stuff to the 'runlocal.lua' file,
since that's included by all the other files, and I'm now using temp
files in several of the test programs.

A few tests have been moved around to where they make more sense.