oocairo: remove unnecessary '_string' functions
The constructor function 'image_surface_create_from_png_string' and the
method 'write_to_png_string' were both unnecessary because the same
thing can be done using the normal versions with a file handle, using
for example my 'memoryfile' module to read and write the data from
memory.  I don't think that's an unreasonable requirement, and this
saves quite a bit of code (more in fact because there are writing
functions which I haven't bound yet).