oocairo: allow writing PDF and SVG to file handle
The 'svg_surface_create' and 'pdf_surface_create' functions now accept
file handles as well as filenames.  To make this work I've had to add
information to the surface userdatas, since the closure struct passed to
the file writing callback has to stay alive for as long as the surface.
I've put the pointer to the actual surface at the start of the struct,
so that most code dereferencing it doesn't need to change (it can just
treat it as a 'cairo_surface_t*').  I keep a reference to the file
handle object in the registry until the surface is garbage collected.

These two functions also now throw an exception deliberately if the file
handle is some random type like bool, rather than waiting for the write
callback to get into trouble.