oocairo: support Cairo where not all features are compiled in
The parts of the C code which rely on particular features (like SVG
surfaces or PNG support) being available in the local copy of Cairo are
now framed with preprocessor stuff to disable them when the feature
isn't available.  Boolean values are provided in the main module table
to indicate which features are available.