oocairo: rename 'runloca.lua' and tidy up examples
The 'runlocal.lua' file is now 'test-setup.lua', and is only used for
the test suite.  I've taken it out of the examples, because it would
serve to be confusing, especially when it doesn't do what it was meant
to on platforms that aren't exactly like mine.  I've also added more
explanation in comments in the examples, and cleaned up the code in a
couple of places.