oocairo: bind new toy font face methods in Cairo 1.8
Added a new constructor function 'Cairo.toy_font_face_create', for
creating a font face object in the same way that the 'select_font_face'
method does, but without setting it on a context object.  To make that
useful also bind the 'set_font_face' object (which I hadn't bound before
because there was no way to create a font face other than setting it on
a context object anyway).

Add methods for getting the family, slant style, and weight of a font face
object, new in Cairo 1.8.

Methods which are only available in Cairo 1.8 are now now defined in
the metatables or the main module table, rather than being defined as
functions which always throw an exception.  This makes it easier to test
if they're available, and more consistent with the way other methods are
handled (ones which are only available when some feature flag is set).
The test suite now tests for the presence of these new 1.8 functions by
seeing if they're defined, rather than checking the Cairo version number
in a dubious way.

Factored out the code for creating fontface userdatas, now that it's used
in two places.