oocairo: preliminary support for user fonts
Added some basic support for the new 'user fonts' feature in Cairo 1.8.
Text clusters aren't handled yet, and scaled font objects aren't passed to
the callbacks, because in neither case do I have a Lua representation for
those objects.

A new example is provided which draws each glyph as a box with indications
of which bits are set in its glyph index.

The user font callbacks in C retrieve the Lua callbacks from a table in
the registry, through a reference held in the userdata of the font face.

Some of the callbacks can provide font or text extents information, which
is copied back from the extents tables into the C structs with new
functions 'from_lua_font_extents' 'from_lua_text_extents'.

Factored out code for creating context userdatas into
'create_context_userdata' since some of the callbacks have to do it.