oocairo: refactor conversion between strings and enums
For some of the new enumerations I'd added functions to translate their
values into Lua strings, and to decode such a string supplied as an argument,
to simplify writing getters and setters.  I now use functions like these
for all the Cairo enumerations I use.  Also, since these functions are all
the same except for the Cairo type name and the names of the constant arrays
listing the values and strings, I've defined them all with macros to reduce
duplication.

The 'antialias_from_lua' function is still defined by hand because it also
accepts boolean values.

A few of the arrays had to be renamed so that the first part of their name
matches the middle part of the Cairo type name, to make the macro simpler.

I've used this technique for converting from 'cairo_surface_type_t' values,
where before they didn't have the arrays defined, since they're only returned
by getters, not accepted by any setters.  The code which used to do the
conversion in this case had lots of '#if' stuff to exclude values which
couldn't occur because the relevant feature wasn't compiled in, but that
was unnecessary, since the enumeration values are defined anyway.

Added some more tests to check that nil values are or aren't accepted in
the right places.

The 'linejoin_values' array (now 'line_join_values') had the wrong type,
but C's type checking is too limited to have noticed.