oocairo: add scaled font class
New type of object that I didn't get around to binding in the first release.
There's a top-level function for creating them from an unscaled font face
object, and getters and setters on context objects.

For some of the tests I've moved utility functions out of other test
programs into 'test-setup.lua' so that they can be used in the scaled
font test program as well.

Also fix a POD bodge in the fontface man page.