oocairo: support drawing glyphs with text clusters
Added the 'text_to_glyphs' method on scaled fonts to figure out which
glyphs to render for a piece of text, and also what the text clusters are.
I've got a simple representation of text clusters in Lua as tables of
two-item arrays.  The text cluster flags are folded in to the same clusters
table with string keys (although there's only one so far in Cairo, for
which I'm using the key 'backward').

Added the 'show_text_glyphs' method on context objects to draw glyphs
with text clusters provided.  Passing in the glyphs and clusters tables
from the 'text_to_glyphs' method works.

Surface objects now have a 'has_show_text_glyphs' method.

All of this only works with Cairo 1.8, so the methods are simply undefined
when compiling with earlier versions.