oocairo: add tests for multiple __gc calls
I have a feeling, although I'm not really sure, that Lua can sometimes
call the '__gc' metamethod more than once for an object, which is why
it's important that it nulls the pointer held in the userdata.  It's
also important that the destruction code doesn't break when there's a
null pointer.  This all seems to work automatically with the Cairo
destroy functions, but it's as well to have tests for it just in case.