oocairo: proper allocation of glyphs and fix leaks
Arrays of glyphs and text clusters need to be allocated with a special
function in Cairo 1.8 in the cases where they're passed back to some of
the new functions which handle the freeing of them, and there are also
places where they have to be freed with a special function.  These new
allocation and freeing functions don't exist in Cairo 1.6, but aren't
needed, so there we just use malloc and free.

In some places, when I passed an array of glyphs to a function I wasn't
freeing it at all, so there were some memory leaks.