oocairo: fix return values from user font text_to_glyphs
Fixed some problems, or at least potential problems, with receiving
glyphs array returned from function, and receive text clusters array
in the same way (if provided).  If glyphs aren't returned then use a
negative 'num_glyphs' value to tell Cairo that we're passing on this,
so it can fall back on other stuff.  Clusters are optional anyway, so
not returning them is OK as well.  A new boolean arg is passed to the
callback indicating whether text clusters will be used if they're returned.