oocairo: basic support for Gtk+ integration
New constructor function 'context_create_for_gdk_window' to create a
Cairo context for drawing on a GdkWindow, and two new context methods
'set_source_pixbuf' and 'set_source_pixmap' for accessing image data
from GtkPixbuf and GdkPixmap objects.  All of this only works if the
Lua-Gnome module 'gtk' is loaded and accessible in the global variables.

Included an example to demonstrate drawing an image loaded through Gdk,
and just for fun some controls for adjusting how it's drawn.  I've also
taken the oppurtinity to learn how to use layouts designed in Glade to
create windows (although I'm using the GtkBuilder interface instead, so
the Glade file is converted to that format).