oocairo: new 'get_data' method for image surfaces
It's now possible to get access to the raw image data from an image
surface, along with the stride number and an indication of the
platform's endianness (both of which are needed to decode the pixels).

Included documentation about the format of the data, which is in some
ways a little more explicit than the actual Cairo documentation, so
not necessarily safe.  Also an example which should (in theory) handle
the different endiannesses.

Original patch by Julien Danjou, although I ended up doing it a bit
differently.