oocairo: add a second example of user fonts
I kept this out of the previous release because it seemed redundant, but
it's a cooler effect than the first example, and now that I've cleaned up
the code a bit and added lots of comments, it's demonstrating a few useful
techniques.  It's particularly interesting because it uses user fonts to
distort another font rather than creating one from scratch.