oocairo: improve examples to ready them for examples web page
I want to publish the descriptions at the start of the example programs
along with their output images on the website, to make it easier for people
to see how to use this module.  To that end I've improved those descriptions,
made various stylistic tweaks to the source, and in a few cases made the
output images a little prettier.  I've also adjusted many of the examples
to produce smaller images, especially trying to reduce the width, so that
they fit better onto a web page.

The 'compositing' example now shows all the operators, and labels them.
The 'fill-rule' and 'arc' examples now draw a stroke as well as filling
their shapes, to show where the path actually goes.  The 'inside-tests'
example now fill and strokes the paths it uses in light grey before drawing
the dots on, so that you can see where the real shapes are.  In a few
examples I indulged my perverse aversion to global variables.  I had to
tweak the 'repeated-patterns' example to make its output filename match
the source code filename, because that's what is expected by the program
I've written to turn this stuff into a web page.  The 'tolerance' example
now draws the curves down the image rather than across, to make it thinner.