oocairo: rename function to 'image_surface_create_from_data'
It was called 'image_surface_create_for_data' to match the native C API,
but actually it's a little different.  Unlike the C version, the data
supplied to this one is a Lua immutable string, and so the actual buffer
used for drawing will be a copy I make of it.  This makes the function
more like 'create_from_png' than something special that allows sharing
buffers with other libraries.  Also, if Cairo ever grows a mutable string
type I might want the other name for a different binding of this.